pyproject.toml
[project]
name = "my-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "google-cloud-bigquery (>=3.41.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
